+++ NEWS.md
@@ -1,6 +1,7 @@
 ## tidydice 0.0.6 (DEV)
 
 * improve vignette
+* add function binom()
 * add function binom_dice()
 * add function binom_coin()
 * add function plot_binom()